requirements.txt
/po-generator-app/
├── app.py               # Main Flask application
├── static/              # Static assets (CSS, JS)
├── templates/           # HTML templates
├── logic/               # Business logic modules
├── uploads/             # Temporary file storage
├── outputs/             # Generated files
└── requirements.txt     # Dependencies

Flask==3.0.2             # Web framework
Flask-Dance==7.0.0       # OAuth integration
Authlib==1.3.0           # Authentication library
SQLAlchemy==2.0.27       # Database ORM
alembic==1.13.1          # Database migrations
pandas==2.2.0            # Data manipulation
openpyxl==3.1.2          # Excel file handling
requests==2.31.0         # HTTP client
python-dotenv==1.0.1     # Environment variable management
pytest==8.0.0            # Testing framework
pytest-mock==3.12.0      # Mocking for tests
black==24.1.1            # Code formatting
flake8==7.0.0            # Code linting
gunicorn==21.2.0         # Production server
openai==1.12.0           # AI integration
celery==5.3.6            # Background tasks
redis==5.0.1             # Cache and message broker
Werkzeug==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.1.0.post1
Flask-RESTful
coverage
